requirements.txt
Django>=4.2.0,<5.0.0
asgiref>=3.7.0
sqlparse>=0.4.4
tzdata>=2025.0
pillow>=11.0.0
mysqlclient>=2.2.0
django-simple-captcha>=0.6.2
pymysql>=1.1.0
redis>=5.0.0 
Faker==37.1.0
numpy==1.26.0
uwsgi>=2.0.20
